ah - Add functionality so that only user specific schedules appear and can be changed

+++ src/main/java/edu/ucsb/courses/controllers/ScheduleController.java
@@ -4,6 +4,8 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import javax.validation.Valid;
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -14,7 +16,10 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.DeleteMapping;
 
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -38,30 +43,40 @@ public class ScheduleController {
     @Autowired
     ScheduleRepository scheduleRepository;
 
-    @GetMapping(value = "/createSchedule", produces = "application/json")
+    @PostMapping(value = "/createSchedule", produces = "application/json")
     public ResponseEntity<String> createSchedule(@RequestParam String name,
                                                  @RequestParam String description,
                                                  @RequestParam String quarter,
-                                                 @RequestParam String userId){
-        Schedule newSched = new Schedule(null, name, description, quarter, userId);
+                                                 @RequestHeader("Authorization") String authorization){
+        DecodedJWT jwt = JWT.decode(authorization.substring(7));
+        Schedule newSched = new Schedule(null, name, description, quarter, jwt.getSubject());
         Schedule savedSched= scheduleRepository.save(newSched);
         return ResponseEntity.ok().body(savedSched.toString());
     }
 
-    @GetMapping(value = "/deleteSchedule", produces = "application/json")
-    public ResponseEntity<String> deleteSchedule(@RequestParam String id){
+    @DeleteMapping(value = "/deleteSchedule", produces = "application/json")
+    public ResponseEntity<String> deleteSchedule(@RequestHeader("Authorization") String authorization, @RequestParam String id){
+        DecodedJWT jwt = JWT.decode(authorization.substring(7));
         Long castId = Long.parseLong(id);
+        Optional<Schedule> target = scheduleRepository.findById(castId);
+        if (!target.isPresent() || !target.get().getUserId().equals(jwt.getSubject())) {
+            return ResponseEntity.notFound().build();
+        }
         scheduleRepository.deleteById(castId);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
     }
 
     @GetMapping(value = "/getSchedule", produces = "application/json")
-    public ResponseEntity<String> getSchedule(@RequestParam String id){
+    public ResponseEntity<String> getSchedule(@RequestHeader("Authorization") String authorization, @RequestParam String id) 
+        throws JsonProcessingException{
+        DecodedJWT jwt = JWT.decode(authorization.substring(7));
         Long castId = Long.parseLong(id);
         Optional<Schedule> target = scheduleRepository.findById(castId);
         if (target.isPresent()) {
-            String body = target.get().toString();
-            return ResponseEntity.ok().body(body);
+            if (target.get().getUserId().equals(jwt.getSubject())){
+                String body = target.get().toString();
+                return ResponseEntity.ok().body(body);
+            }
         }
         return ResponseEntity.badRequest().build();
     }
